Move App component file
