+++ lib/main.js
@@ -1,5 +1,5 @@
 import Vue from 'vue'
-import App from './App.vue'
+import App from './components/app.vue'
 
 new Vue({ // eslint-disable-line no-new
   el: '#app',
